requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pandas==2.2.3
spacy==3.8.2
ollama==0.4.4
python-multipart==0.0.12
pytesseract==0.3.13
Pillow==10.4.0
pdfplumber==0.11.4
numpy==1.26.0
scikit-learn==1.5.0
PyJWT==2.8.0
bcrypt==4.1.2
openpyxl==3.1.2
